errno for retry

+++ rootfs/lib/src/ipc.c
@@ -112,7 +112,7 @@ package_t* ipc_req(int pid, uint32_t buf_size, uint32_t type, void* data, uint32
 		return NULL;
 
 	int i = ipc_send(id, type, data, size);
-	if(i != size || !reply) {
+	if(i != (int32_t)size || !reply) {
 		ipc_close(id);
 		return NULL;
 	}
